Add StrategySettings
* Move Strategy class from Bet.py into separate Strategy.py
