+++ TwitchChannelPointsMiner/classes/entities/Bet.py
@@ -1,241 +1,14 @@
 import copy
-import logging
 from enum import Enum, auto
-from random import uniform
 
 from millify import millify
 
-from TwitchChannelPointsMiner.classes.Settings import Settings
-from TwitchChannelPointsMiner.utils import char_decision_as_index, float_round
-
-logger = logging.getLogger(__name__)
-
-
-class Strategy(object):
-    MOST_VOTED = "MostVoted"
-    HIGH_ODDS = "HighOdds"
-    SMART_HIGH_ODDS = "SmartHighOdds"
-    PERCENTAGE = "Percentage"
-    SMART = "Smart"
-
-    def __init__(self, outcomes: list, settings: object):
-        self.outcomes = outcomes
-        self.decision: dict = {}
-        self.settings = settings
-
-    def get_instance(self):
-        subclass = globals()[self.settings.strategy]
-        return subclass(self.outcomes, self.settings)
-
-    def return_choice(self, key) -> str:
-        return "A" if self.outcomes[0][key] > self.outcomes[1][key] else "B"
-
-    def calculate_before(self):
-        self.decision = {"choice": None, "amount": 0, "id": None}
-
-    def calculate_middle(self):
-        pass
-
-    def calculate_after(self, balance: int):
-        if self.settings.only_doubt:
-            self.decision["choice"] = "B"
-        if self.decision["choice"] is not None:
-            index = char_decision_as_index(self.decision["choice"])
-            self.decision["id"] = self.outcomes[index]["id"]
-            amounts = [
-                self.decision["amount"],
-                int(balance * (self.settings.percentage / 100)),
-                self.settings.max_points,
-            ]
-            self.decision["amount"] = min(x for x in amounts if x != 0)
-            if (
-                self.settings.stealth_mode is True
-                and self.decision["amount"]
-                >= self.outcomes[index][OutcomeKeys.TOP_POINTS]
-            ):
-                reduce_amount = uniform(1, 5)
-                self.decision["amount"] = (
-                    self.outcomes[index][OutcomeKeys.TOP_POINTS] - reduce_amount
-                )
-            self.decision["amount"] = int(self.decision["amount"])
-
-    def calculate(self, balance: int) -> dict:
-        self.calculate_before()
-        self.calculate_middle()
-        self.calculate_after(balance)
-        return self.decision
-
-    def skip_before(self):
-        if self.settings.always_bet:
-            self.log_skip("Odd is too low, but always_bet activated")
-            return False, 0
-
-    def skip_middle(self):
-        pass
-
-    def skip_after(self):
-        if self.settings.filter_condition is not None:
-            self.decision.setdefault("choice", None)  # for tests
-            # key == by , condition == where
-            key = self.settings.filter_condition.by
-            condition = self.settings.filter_condition.where
-            value = self.settings.filter_condition.value
-
-            fixed_key = (
-                key
-                if key not in [OutcomeKeys.DECISION_USERS, OutcomeKeys.DECISION_POINTS]
-                else key.replace("decision", "total")
-            )
-            if key in [OutcomeKeys.TOTAL_USERS, OutcomeKeys.TOTAL_POINTS]:
-                compared_value = (
-                    self.outcomes[0][fixed_key] + self.outcomes[1][fixed_key]
-                )
-            else:
-                outcome_index = char_decision_as_index(self.decision["choice"])
-                compared_value = self.outcomes[outcome_index][fixed_key]
-
-            # Check if condition is satisfied
-            if condition == Condition.GT:
-                if compared_value > value:
-                    return False, compared_value
-            elif condition == Condition.LT:
-                if compared_value < value:
-                    return False, compared_value
-            elif condition == Condition.GTE:
-                if compared_value >= value:
-                    return False, compared_value
-            elif condition == Condition.LTE:
-                if compared_value <= value:
-                    return False, compared_value
-            return True, compared_value  # Else skip the bet
-        else:
-            return False, 0  # Default don't skip the bet
-
-    def skip(self) -> bool:
-        skip_results = [self.skip_before(), self.skip_middle(), self.skip_after()]
-        return next(item for item in skip_results if item is not None)
-
-    def __str__(self):
-        return self.name
-
-
-class SmartHighOdds(Strategy):
-    def both_odds_too_low(self) -> bool:
-        return (
-            self.outcomes[0][OutcomeKeys.ODDS] <= self.settings.target_odd
-            and self.outcomes[1][OutcomeKeys.ODDS] <= self.settings.target_odd
-        )
-
-    def is_only_doubt(self) -> bool:
-        return (
-            self.outcomes[1][OutcomeKeys.ODDS] <= self.settings.target_odd
-            and self.settings.only_doubt
-        )
-
-    def log_skip(self, string) -> str:
-        logger.info(
-            string,
-            extra={
-                "emoji": ":pushpin:",
-                "color": Settings.logger.color_palette.BET_GENERAL,
-            },
-        )
-
-    def skip_middle(self):
-        if (
-            self.outcomes[0][OutcomeKeys.TOTAL_POINTS] > 0
-            and self.outcomes[1][OutcomeKeys.TOTAL_POINTS] == 0
-        ):
-            self.log_skip("No bet on B")
-            return False, 0
-        if (
-            self.outcomes[0][OutcomeKeys.TOTAL_POINTS] == 0
-            and self.outcomes[1][OutcomeKeys.TOTAL_POINTS] > 0
-        ):
-            if not self.settings.only_doubt:
-                self.log_skip("No bet on A")
-                return False, 0
-
-        if self.both_odds_too_low() or self.is_only_doubt():
-            if self.both_odds_too_low():
-                self.log_skip("Odd is too low")
-            elif self.is_only_doubt():
-                self.log_skip("Odd is too low and only_doubt activated")
-            self.log_skip(f"{self.get_outcome(0)}; {self.get_outcome(1)}")
-            self.log_skip(f"Target odd: {self.settings.target_odd}")
-            return True, 0  # Skip
-
-    def calculate_sho_bet(self, index):
-        low_odd_points = self.outcomes[1 - index][OutcomeKeys.TOTAL_POINTS]
-        high_odd_points = self.outcomes[index][OutcomeKeys.TOTAL_POINTS]
-        if self.both_odds_too_low() or self.is_only_doubt():
-            return 10
-        elif high_odd_points <= 50:  # in case no one bet
-            return 50
-        else:
-            target_odd = self.settings.target_odd
-            if self.outcomes[index][OutcomeKeys.ODDS] > (target_odd * 2):
-                # don't bet too much if odds is too high
-                target_odd = self.outcomes[index][OutcomeKeys.ODDS] / 2
-            return int((low_odd_points / (target_odd - 1)) - high_odd_points)
-
-    def calculate_middle(self):
-        self.decision["choice"] = self.return_choice(OutcomeKeys.ODDS)
-        if self.decision["choice"] is not None:
-            index = char_decision_as_index(self.decision["choice"])
-            self.decision["amount"] = int(self.calculate_sho_bet(index))
-
-
-class MostVoted(Strategy):
-    def calculate_middle(self):
-        self.decision["choice"] = self.return_choice(OutcomeKeys.TOTAL_USERS)
-
-
-class HighOdds(Strategy):
-    def calculate_middle(self):
-        self.decision["choice"] = self.return_choice(OutcomeKeys.ODDS)
-
-
-class Percentage(Strategy):
-    def calculate_middle(self):
-        self.decision["choice"] = self.return_choice(OutcomeKeys.ODDS_PERCENTAGE)
-
-
-class Smart(Strategy):
-    def calculate_middle(self):
-        difference = abs(
-            self.outcomes[0][OutcomeKeys.PERCENTAGE_USERS]
-            - self.outcomes[1][OutcomeKeys.PERCENTAGE_USERS]
-        )
-        self.decision["choice"] = (
-            self.return_choice(OutcomeKeys.ODDS)
-            if difference < self.settings.percentage_gap
-            else self.return_choice(OutcomeKeys.TOTAL_USERS)
-        )
-
-
-class Condition(Enum):
-    GT = auto()
-    LT = auto()
-    GTE = auto()
-    LTE = auto()
-
-    def __str__(self):
-        return self.name
-
-
-class OutcomeKeys(object):
-    # Real key on Bet dict ['']
-    PERCENTAGE_USERS = "percentage_users"
-    ODDS_PERCENTAGE = "odds_percentage"
-    ODDS = "odds"
-    TOP_POINTS = "top_points"
-    # Real key on Bet dict [''] - Sum()
-    TOTAL_USERS = "total_users"
-    TOTAL_POINTS = "total_points"
-    # This key does not exist
-    DECISION_USERS = "decision_users"
-    DECISION_POINTS = "decision_points"
+from TwitchChannelPointsMiner.classes.entities.Strategy import (
+    OutcomeKeys,
+    Strategy,
+    StrategySettings,
+)
+from TwitchChannelPointsMiner.utils import float_round
 
 
 class DelayMode(Enum):
@@ -267,61 +40,63 @@ class BetSettings(object):
     __slots__ = [
         "strategy",
         "percentage",
-        "percentage_gap",
         "max_points",
-        "target_odd",
         "only_doubt",
-        "always_bet",
         "minimum_points",
         "stealth_mode",
         "filter_condition",
         "delay",
         "delay_mode",
+        "strategy_settings",
     ]
 
     def __init__(
         self,
         strategy: Strategy = None,
         percentage: int = None,
-        percentage_gap: int = None,
         max_points: int = None,
-        target_odd: float = None,
         only_doubt: bool = None,
-        always_bet: bool = None,
         minimum_points: int = None,
         stealth_mode: bool = None,
         filter_condition: FilterCondition = None,
         delay: float = None,
         delay_mode: DelayMode = None,
+        strategy_settings: StrategySettings = None,
     ):
         self.strategy = strategy
         self.percentage = percentage
-        self.percentage_gap = percentage_gap
         self.max_points = max_points
-        self.target_odd = target_odd
         self.only_doubt = only_doubt
-        self.always_bet = always_bet
         self.minimum_points = minimum_points
         self.stealth_mode = stealth_mode
         self.filter_condition = filter_condition
         self.delay = delay
         self.delay_mode = delay_mode
+        self.strategy_settings = StrategySettings(
+            strategy=strategy, **strategy_settings
+        ).get_instance()
 
     def default(self):
-        self.strategy = self.strategy if not None else Strategy.SMART
-        self.percentage = self.percentage if not None else 5
-        self.percentage_gap = self.percentage_gap if not None else 20
-        self.max_points = self.max_points if not None else 50000
-        self.target_odd = self.target_odd if not None else 3
-        self.only_doubt = self.only_doubt if not None else False
-        self.always_bet = self.always_bet if not None else True
-        self.minimum_points = self.minimum_points if not None else 0
-        self.stealth_mode = self.stealth_mode if not None else False
-        self.delay = self.delay if not None else 6
-        self.delay_mode = self.delay_mode if not None else DelayMode.FROM_END
+        self.strategy = self.strategy if self.strategy is not None else Strategy.SMART
+        self.percentage = self.percentage if self.percentage is not None else 5
+        self.max_points = self.max_points if self.max_points is not None else 50000
+        self.only_doubt = self.only_doubt if self.only_doubt is not None else False
+        self.minimum_points = (
+            self.minimum_points if self.minimum_points is not None else 0
+        )
+        self.stealth_mode = (
+            self.stealth_mode if self.stealth_mode is not None else False
+        )
+        self.delay = self.delay if self.delay is not None else 6
+        self.delay_mode = (
+            self.delay_mode if self.delay_mode is not None else DelayMode.FROM_END
+        )
+        self.strategy_settings = (
+            self.strategy_settings if self.strategy_settings is not None else {}
+        )
 
     def __repr__(self):
-        return f"BetSettings(strategy={self.strategy}, percentage={self.percentage}, percentage_gap={self.percentage_gap}, max_points={self.max_points}, minimum_points={self.minimum_points}, stealth_mode={self.stealth_mode})"
+        return f"BetSettings(strategy={self.strategy}, percentage={self.percentage}, max_points={self.max_points}, minimum_points={self.minimum_points}, stealth_mode={self.stealth_mode})"
 
 
 class Bet(object):
@@ -425,4 +200,7 @@ class Bet(object):
         return Strategy(self.outcomes, self.settings).get_instance().skip()
 
     def calculate(self, balance: int) -> dict:
-        return Strategy(self.outcomes, self.settings).get_instance().calculate(balance)
+        self.decision = (
+            Strategy(self.outcomes, self.settings).get_instance().calculate(balance)
+        )
+        return self.decision
